// src/api/generated/models/OpportunityResponse.ts
export interface OpportunityResponse {
    /**
     * 
     * @type {string}
     * @memberof OpportunityResponse
     */
    opportunityId?: string;
    /**
     * 
     * @type {string}
     * @memberof OpportunityResponse
     */
    opportunityName?: string;
    /**
     * 
     * @type {string}
     * @memberof OpportunityResponse
     */
    customerId?: string;
    /**
     * 
     * @type {string}
     * @memberof OpportunityResponse
     */
    customerName?: string;
    /**
     * 
     * @type {string}
     * @memberof OpportunityResponse
     */
    contactId?: string;
    /**
     * 
     * @type {string}
     * @memberof OpportunityResponse
     */
    contactName?: string;
    /**
     * 
     * @type {string}
     * @memberof OpportunityResponse
     */
    stageId?: string;
    /**
     * 
     * @type {string}
     * @memberof OpportunityResponse
     */
    stageName?: string;
    /**
     * 
     * @type {number}
     * @memberof OpportunityResponse
     */
    amount?: number;
    /**
     * 
     * @type {number}
     * @memberof OpportunityResponse
     */
    probability?: number;
    /**
     * 
     * @type {Date}
     * @memberof OpportunityResponse
     */
    expectedCloseDate?: Date;
    /**
     * 
     * @type {Date}
     * @memberof OpportunityResponse
     */
    actualCloseDate?: Date;
    /**
     * 
     * @type {string}
     * @memberof OpportunityResponse
     */
    leadSource?: string;
    /**
     * 
     * @type {string}
     * @memberof OpportunityResponse
     */
    description?: string;
    /**
     * 
     * @type {string}
     * @memberof OpportunityResponse
     */
    nextStep?: string;
    /**
     * 
     * @type {string}
     * @memberof OpportunityResponse
     */
    competitorInfo?: string;
    /**
     * 
     * @type {string}
     * @memberof OpportunityResponse
     */
    assignedUserId?: string;
    /**
     * 
     * @type {string}
     * @memberof OpportunityResponse
     */
    assignedUserName?: string;
    /**
     * 
     * @type {string}
     * @memberof OpportunityResponse
     */
    branchId?: string;
    /**
     * 
     * @type {string}
     * @memberof OpportunityResponse
     */
    branchName?: string;
    /**
     * 
     * @type {string}
     * @memberof OpportunityResponse
     */
    wonReason?: string;
    /**
     * 
     * @type {string}
     * @memberof OpportunityResponse
     */
    lostReason?: string;
    /**
     * 
     * @type {string}
     * @memberof OpportunityResponse
     */
    tenantId?: string;
    /**
     * 
     * @type {string}
     * @memberof OpportunityResponse
     */
    useAt?: string;
    /**
     * 
     * @type {Date}
     * @memberof OpportunityResponse
     */
    createdDate?: Date;
    /**
     * 
     * @type {Date}
     * @memberof OpportunityResponse
     */
    updatedDate?: Date;
}